Rearrange webhook parsing and event handling
